add a little error handling behing the getting of data from the api
